Refactored the request functionality up into the abstract class.
