+++ src/main/java/com/stehno/ersatz/Expectations.java
@@ -1,12 +1,12 @@
 /**
  * Copyright (C) 2016 Christopher J. Stehno
- *
+ * <p>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -20,5 +20,5 @@ package com.stehno.ersatz;
  */
 public interface Expectations {
 
-    GetRequest get(String path);
+    Request get(String path);
 }
