+++ src/main/java/com/stehno/ersatz/model/ExpectationsImpl.java
@@ -1,12 +1,12 @@
 /**
  * Copyright (C) 2016 Christopher J. Stehno
- *
+ * <p>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -16,7 +16,6 @@
 package com.stehno.ersatz.model;
 
 import com.stehno.ersatz.Expectations;
-import com.stehno.ersatz.GetRequest;
 import com.stehno.ersatz.Request;
 import io.undertow.server.HttpServerExchange;
 
@@ -31,8 +30,8 @@ public class ExpectationsImpl implements Expectations {
 
     private final List<Request> requests = new ArrayList<>();
 
-    public GetRequest get(final String path) {
-        GetRequest request = new GetRequestImpl(path);
+    public Request get(final String path) {
+        Request request = new GetRequestImpl(path);
         requests.add(request);
         return request;
     }
